+++ app/melos_template/lib/core/network/api_client/weather_api_client.dart
@@ -10,7 +10,6 @@ part 'weather_api_client.g.dart';
 
 @riverpod
 WeatherApiClient weatherApiClient(Ref ref) {
-
   final dio = ref.read(dioProvider(isRequireAuthenticate: false));
   return WeatherApiClient(dio, baseUrl: ApiEndpoint.openWeather);
 }
